Introduction of QProxy: QVariant UD pointers
Binding quizzes to QTreeWidgetItems no longer causes a deep copy of the
quiz, but is instead wrapped by a QProxy pointer container, which gets
copied, so we're good to go.

It's not the cleanest solution because the instantiated template class
of QProxy with your custom user-data pointer type still needs to be
registered with Qt's META_TYPE registry, something like this:

typedef QProxy<Canvas::MyTYpe> PMyType;
Q_DECLARE_METATYPE(PMyType)

And if you're using the type in signals/slots:

qRegisterMetaType<PMyType>("PMyType");

Take a look at include/type_exports.hpp for more information.